akv-sync-python/requirements.txt
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
